+++ src/Components/ScrollToButton/ScrollToButton.component.scss
@@ -1,0 +1,5 @@
+.go-down-arrow {
+  font-size: 1.5em;
+  position: absolute;
+  bottom: 8%;
+}
