+++ src/Components/WelcomePage/WelcomePage.component.scss
@@ -18,8 +18,4 @@
   justify-content: center;
 }
 
-.go-down-arrow {
-  font-size: 1.5em;
-  position: absolute;
-  bottom: 8%;
-}
+
